refactored Contact info
